Update ContactFormTwo.vue component and adjust indentation
The main changes in this commit include a modification to the 'businessType' form field validation schema to accommodate object structure, and the adjustment of indentation within the file for better readability. Prior to this commit, 'businessType' only accepted a string value, however, now it uses object structure which accepts both label and value properties.